requirements.txt
Flask==3.1.1

Flask-Cors==6.0.1

Flask-Migrate==4.1.0

Flask-SQLAlchemy==3.1.1

SQLAlchemy==2.0.40

psycopg2-binary==2.9.10

certifi==2026.1.4

charset-normalizer==3.4.4

colorama==0.4.6

filelock==3.20.3

fsspec==2026.1.0

huggingface-hub==0.36.0

idna==3.11

Jinja2==3.1.6

joblib==1.5.3

MarkupSafe==3.0.3

mpmath==1.3.0

networkx==3.6.1

numpy==2.4.1

packaging==26.0

PyMuPDF==1.26.7

PyYAML==6.0.3

regex==2026.1.15

requests==2.32.5

safetensors==0.7.0

scikit-learn==1.8.0

scipy==1.17.0

sentence-transformers==5.2.0

chromadb==0.4.14

pysqlite3-binary==0.5.4

sympy==1.14.0

threadpoolctl==3.6.0

tokenizers==0.22.2

torch==2.10.0

tqdm==4.67.1

transformers==4.57.6

typing_extensions==4.15.0

urllib3==2.6.3

google-genai>=1.0.0

python-dotenv>=1.0.1

gunicorn==23.0.0

